cloud_functions/auto_sync/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-pubsub==2.*
supabase==2.*
python-dotenv==1.*
